refactor: 차량 폼 입력 및 유효성 검증 로직 커스텀 훅으로 분리
